Just rename PagedKVCache
