+++ sharktank/sharktank/export_layer/export_kv_cache.py
@@ -11,7 +11,7 @@ from iree.turbine.aot import *
 
 from sharktank.types import SplitPrimitiveTensor
 from sharktank.ops import reshard_split, replicate
-from sharktank.layers.kv_cache import PagedKVCache
+from sharktank.layers.kv_cache import PagedAttention
 from ..utils import cli
 
 
@@ -60,7 +60,7 @@ def main():
     page_count = bs * seq_length // block_seq_stride
     write_seq_length = seq_length - 4
 
-    cache = PagedKVCache(
+    cache = PagedAttention(
         block_seq_stride=block_seq_stride,
         transformer_block_count=transformer_block_count,
         attn_head_count=attn_head_count,
