--- sharktank/sharktank/ops/paged_attention.py
@@ -1,344 +1,0 @@
-# Copyright 2024 Advanced Micro Devices, Inc.
-#
-# Licensed under the Apache License v2.0 with LLVM Exceptions.
-# See https://llvm.org/LICENSE.txt for license information.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-"""Light-weight encapsulations for various forms of attention KV-caches.
-
-These are not complete abstractions: they are primarily focused on making
-tightly coupled transformer blocks a bit less "stringy" with loose tensors
-and dims floating around everywhere.
-"""
-
-from typing import Optional, Union, List, Sequence
-
-import math
-
-import torch
-
-from ..types import SplitPrimitiveTensor, ReplicatedTensor
-from .. import ops
-
-__all__ = ["PagedAttention"]
-
-
-class PagedAttention:
-    def get_page_size(self):
-        return (
-            self.cache_partition_count
-            * self.block_seq_stride
-            * self.attn_head_count
-            * self.attn_head_dim
-            // self.shard_count
-        )
-
-    def __init__(
-        self,
-        *,
-        transformer_block_count: int,
-        attn_head_count: int,
-        attn_head_dim: int,
-        cache_partition_count: int = 2,
-        block_seq_stride: int = 16,
-        dtype: torch.dtype = torch.float32,
-        device: Optional[torch.device] = None,
-        shard_count: int = 1,
-    ):
-        self.transformer_block_count = transformer_block_count
-        self.attn_head_count = attn_head_count
-        self.attn_head_dim = attn_head_dim
-        self.cache_partition_count = cache_partition_count
-        self.block_seq_stride = block_seq_stride
-        self.shard_count = shard_count
-        if attn_head_count % shard_count != 0:
-            raise ValueError(
-                f"The attention head count {attn_head_count} must be a multiple of the tensor parallelism size {shard_count}."
-            )
-
-        # Some derived values based on attributes.
-        self.sub_page_dims = [
-            self.transformer_block_count,
-            self.cache_partition_count,
-            self.block_seq_stride,
-            self.attn_head_count // self.shard_count,
-            self.attn_head_dim,
-        ]
-        self.page_slab_flat_dim = math.prod(self.sub_page_dims)
-        self.device = device
-        self.dtype = dtype
-
-    def unflatten_page_table(
-        self, state: Sequence[Union[torch.Tensor, SplitPrimitiveTensor]]
-    ) -> Union[torch.Tensor, SplitPrimitiveTensor]:
-        """Unflattens the 2D page table to a 6D tensor."""
-        assert len(state) == 1, f"Expected 1-element state. Got: {len(state)}"
-        page_slab = state[0]
-        if self.shard_count == 1:
-            assert not isinstance(page_slab, SplitPrimitiveTensor)
-            return page_slab.unflatten(1, self.sub_page_dims)
-        else:
-            assert isinstance(page_slab, SplitPrimitiveTensor)
-            assert self.shard_count == page_slab.shard_count
-            shards = [
-                shard.unflatten(1, self.sub_page_dims) for shard in page_slab.shards
-            ]
-            return SplitPrimitiveTensor(ts=shards, shard_dim=4)
-
-    def shard_state(
-        self, state: List[torch.Tensor]
-    ) -> Sequence[Union[torch.Tensor, SplitPrimitiveTensor]]:
-        """Shard an unsharded state.
-        We can't just split the slab on the sub page dims.
-        First it needs to be reinterpreted into the actual shape.
-        The split the head dimension, then flatten each shard.
-        This is a work-around for the lack of block-cyclic sharded tensor type."""
-        if self.shard_count == 1:
-            return state
-
-        page_table = state[0].reshape(
-            [
-                -1,
-                self.transformer_block_count,
-                self.cache_partition_count,
-                self.block_seq_stride,
-                self.attn_head_count,
-                self.attn_head_dim,
-            ]
-        )
-        sharded_page_table = ops.reshard_split(
-            page_table, dim=4, count=self.shard_count
-        )
-        shards = [
-            ops.flatten(shard, start_dim=1) for shard in sharded_page_table.shards
-        ]
-        flat_sharded_page_table = SplitPrimitiveTensor(ts=shards, shard_dim=1)
-        return [flat_sharded_page_table]
-
-    @property
-    def pad_sequence_stride(self) -> int:
-        return self.block_seq_stride
-
-    def read(
-        self,
-        state: Sequence[Union[torch.Tensor, SplitPrimitiveTensor]],
-        *,
-        transformer_block_index: int,
-        seq_len: int,
-        page_ids: Union[torch.Tensor, ReplicatedTensor],
-    ):
-        """Reads K/V caches the page table for the given page_ids.
-
-        Args:
-        state: State struct as returned from allocate().
-        transformer_block_index: The index of the transformer block accessing
-            the cache.
-        page_ids: Tensor of [bs, max_seqlen // block_pos_stride] of page ids
-            to access.
-
-        Returns the K/V cache partitions, linearized. Note that this reference
-        approach to reading by materializing linearly may not be terribly
-        efficient unless if the compiler can fuse the gather.
-        """
-        page_table = self.unflatten_page_table(state)  # 6D
-
-        bs, block_seq_len, *_ = page_ids.shape
-        # Blocks dim 1,2 according to the configured block stride.
-        blocked_shape = [
-            bs,
-            block_seq_len,
-            self.cache_partition_count,
-            self.block_seq_stride,
-            self.attn_head_count // self.shard_count,
-            self.attn_head_dim,
-        ]
-
-        # Gather both partitions and split post gather. This is more
-        # computationally efficient without gather fusion:
-        subblock_table = page_table.flatten(start_dim=0, end_dim=1)
-        page_stride = self.transformer_block_count
-
-        transformer_block_indices = torch.full(
-            (bs, block_seq_len), transformer_block_index
-        )
-        subblock_ids = page_ids * page_stride + transformer_block_indices
-        selected = ops.index_select(subblock_table, 0, subblock_ids.flatten(0, 1))
-
-        selected = selected.unflatten(0, blocked_shape[:2])
-        key = selected[:, :, 0, :seq_len].flatten(1, 2)[:, :seq_len]
-        value = selected[:, :, 1, :seq_len].flatten(1, 2)[:, :seq_len]
-
-        return key, value
-
-    def write_decode(
-        self,
-        state: Sequence[Union[torch.Tensor, SplitPrimitiveTensor]],
-        # List of [bs, 1, attn_head_count, attn_head_dim]
-        cache_partitions: Sequence[Union[torch.Tensor, SplitPrimitiveTensor]],
-        *,
-        transformer_block_index: int,
-        # [bs]
-        seq_positions: Union[torch.Tensor, ReplicatedTensor],
-        # [bs, max_seqlen // block_pos_stride]
-        page_ids: Union[torch.Tensor, ReplicatedTensor],
-    ):
-        """Writes a single batched timestep across all cache partitions.
-
-        Note that this internally loops over the batch size, which cannot be
-        dynamic.
-        """
-        assert all((part.shape[1] == 1 for part in cache_partitions))
-        assert len(cache_partitions) == self.cache_partition_count
-
-        device = self.device
-        page_table = self.unflatten_page_table(state)  # 6D
-        page_table = page_table.flatten(0, 3)
-        bs, *_ = seq_positions.shape
-
-        # [bs, 1, atten_head_count, attn_head_dim]
-        for idx, cache_partition in enumerate(cache_partitions):
-            # [bs, 1]
-            page_index = seq_positions // self.block_seq_stride
-
-            page_id = ops.gather(page_ids, dim=1, index=page_index.unsqueeze(1))
-            page_offset = (seq_positions % self.block_seq_stride).unsqueeze(1)
-
-            # [1, 1]
-            if isinstance(seq_positions, ReplicatedTensor):
-                partitions = [
-                    torch.tensor(idx).unsqueeze(0)
-                    for _ in range(seq_positions.shard_count)
-                ]
-
-                transformer_block = [
-                    torch.full((bs, 1), transformer_block_index, device=device)
-                    for _ in range(seq_positions.shard_count)
-                ]
-
-                partitions = ReplicatedTensor(ts=partitions)
-                transformer_block = ReplicatedTensor(ts=transformer_block)
-            else:
-                partitions = torch.tensor(idx).unsqueeze(0)
-                transformer_block = torch.full(
-                    (bs, 1), transformer_block_index, device=device
-                )
-
-            partitions = partitions.repeat(bs, 1)
-
-            index = page_id
-            index = index * self.transformer_block_count + transformer_block
-            index = index * self.cache_partition_count + partitions
-            index = index * self.block_seq_stride + page_offset
-            values = ops.to(cache_partition, dtype=page_table.dtype)
-            if page_table.dtype == torch.float8_e4m3fnuz:
-                # Workaround for Torch not supporting torch.Tensor.index_copy_ for f8.
-                page_table_as_int8 = page_table.view(dtype=torch.int8)
-                values_int8 = values.view(dtype=torch.int8)
-                page_table_as_int8.index_put_(indices=(index,), values=values_int8)
-            else:
-                page_table.index_put_(indices=(index,), values=values)
-
-        return
-
-    def write_prefill(
-        self,
-        state: Sequence[Union[torch.Tensor, SplitPrimitiveTensor]],
-        cache_partitions: Sequence[Union[torch.Tensor, SplitPrimitiveTensor]],
-        *,
-        transformer_block_index: int,
-        page_ids: Union[torch.Tensor, ReplicatedTensor],
-    ):
-        """Writes cache partitions from a linear layout to the page table.
-
-        This is the inverse of the linear read. The same caveat applies if the
-        in-place scatter cannot be fused.
-        """
-        page_table = self.unflatten_page_table(state)  # 6D
-
-        bs, block_seq_len, *_ = page_ids.shape
-
-        # Reshape the page cache into sub-blocks so that we can index at the
-        # granularity of the transformer_block and cache partition.
-        # This requires us to recompute indices to the sub-block reference
-        # frame.
-        # The subblock slab is organized as:
-        #   [page, attn_layer, cache_partition]
-        # Where the cache line can be 0 (k) or 1 (v).
-        subblock_table = page_table.flatten(start_dim=0, end_dim=2)
-        page_stride = self.transformer_block_count * self.cache_partition_count
-        transformer_block_stride = self.cache_partition_count
-        base_subblock_ids = page_ids * page_stride + (
-            transformer_block_index * transformer_block_stride
-        )
-
-        for index, partition in enumerate(cache_partitions):
-            part_block_view = partition.unflatten(
-                1, (block_seq_len, self.block_seq_stride)
-            )
-            part_block_view = part_block_view.flatten(0, 1)
-
-            subblock_ids = (
-                (base_subblock_ids + index) if index > 0 else base_subblock_ids
-            ).flatten(0, 1)
-
-            part_block = ops.to(part_block_view, dtype=subblock_table.dtype)
-            if subblock_table.dtype == torch.float8_e4m3fnuz:
-                # Workaround for Torch not supporting torch.Tensor.index_copy_ for f8.
-                subblock_table_as_int8 = subblock_table.view(dtype=torch.int8)
-                part_block_as_int8 = part_block.view(dtype=torch.int8)
-                subblock_table_as_int8.index_copy_(0, subblock_ids, part_block_as_int8)
-            else:
-                subblock_table.index_copy_(0, subblock_ids, part_block)
-
-    def attention(
-        self,
-        *,
-        q: torch.Tensor,
-        k: torch.Tensor,
-        v: torch.Tensor,
-        attention_kernel: str,
-        softcap: Optional[float] = None,
-        scale: Optional[torch.Tensor] = None,
-        mask: Optional[torch.Tensor] = None,
-    ):
-        # Non-decomposed
-        if attention_kernel != "decomposed":
-            if softcap is not None:
-                raise ValueError("softcap not supported yet")
-
-            return ops.scaled_dot_product_attention(
-                q=q,  # [bs, ..., sl, dim]
-                k=k,  # [bs, ..., sl, dim]
-                v=v,  # [bs, ..., sl, dim]
-                a=mask,  # [bs, ..., sl, sl]
-                is_causal=mask is None,  # assumes causal masking when true
-                scale=None,  # defaults to 1/sqrt(dim)
-            )
-
-        # Decomposed
-        if attention_kernel == "decomposed":
-            attn_weights = ops.matmul(q, k.transpose(2, 3))
-            if scale is None:
-                attn_weights = attn_weights / math.sqrt(self.attn_head_dim)
-            else:
-                attn_weights = attn_weights * scale
-
-            # Flash attention.
-            if softcap is not None:
-                attn_weights = softcap * torch.tanh(attn_weights / softcap)
-
-            # Apply attention mask.
-            if mask is None:
-                mask = torch.full(
-                    (attn_weights.shape[2], attn_weights.shape[3]), float("-inf")
-                )
-                mask = torch.triu(mask, diagonal=1)[None, None, :, :]
-                attn_weights = attn_weights + mask
-            else:
-                attn_weights = attn_weights + mask
-
-            attn_weights = ops.softmax(
-                ops.to(attn_weights, dtype=torch.float32), dim=-1
-            )
-            attn_weights = ops.to(attn_weights, dtype=q.dtype)
-            return ops.matmul(attn_weights, v)  # (bs, heads, slen, head_dim)
