+++ sharktank/sharktank/utils/create_cache.py
@@ -7,13 +7,13 @@
 from ..layers import *
 
 
-def create_paged_kv_cache(config: LlamaModelConfig) -> PagedKVCache:
+def create_paged_kv_cache(config: LlamaModelConfig) -> PagedAttention:
     if config.kv_cache_type != "paged":
         raise ValueError("Model does not use paged kv cache, cannot create kv cache")
 
     hp = config.hp
     dtype = config.kv_cache_dtype or config.attention_dtype
-    return PagedKVCache(
+    return PagedAttention(
         transformer_block_count=hp.block_count,
         attn_head_count=hp.attention_head_count_kv,
         attn_head_dim=hp.attn_head_dim,
