+++ sharktank/tests/layers/kv_cache_test.py
@@ -28,7 +28,7 @@ def test_paged(dtype: torch.dtype):
     attn_head_dim = 16
     transformer_block_count = 4
     block_seq_stride = 4
-    cache = PagedKVCache(
+    cache = PagedAttention(
         block_seq_stride=block_seq_stride,
         transformer_block_count=transformer_block_count,
         attn_head_count=attn_head_count,
@@ -136,7 +136,7 @@ def test_sharded_paged():
     transformer_block_count = 4
     block_seq_stride = 4
     shard_count = 4
-    cache = PagedKVCache(
+    cache = PagedAttention(
         block_seq_stride=block_seq_stride,
         transformer_block_count=transformer_block_count,
         attn_head_count=attn_head_count,
