+++ sharktank/tests/models/llama/kv_cache_test.py
@@ -11,7 +11,7 @@ import torch.nn as nn
 from sharktank.models.llama.llama import (
     LlamaAttentionBlock,
     PagedLlamaAttentionBlock,
-    PagedKVCache,
+    PagedAttention,
 )
 from sharktank.models.llama.testing import *
 from sharktank.layers.rotary_embedding import RotaryEmbeddingLayer
@@ -41,7 +41,7 @@ class KVCacheTest(unittest.TestCase):
             ffn_dim=self.ffn_dim,
             dtype=self.attention_dtype,
         )
-        self.paged_kv_cache = PagedKVCache(
+        self.paged_kv_cache = PagedAttention(
             transformer_block_count=self.head_count,
             attn_head_count=self.head_count,
             attn_head_dim=self.head_dim,
