+++ sharktank/tests/models/llama/test_llama.py
@@ -29,9 +29,12 @@ def test_llama():
     ids = torch.asarray([ids], dtype=torch.int64)
     block_ids = torch.asarray([[i for i in range(blocks)]]).to(torch.int64)
 
-    cache_state = model.cache.allocate(
-        page_count=config.hp.context_length // config.block_seq_stride
-    )
+    # cache_state = model.cache.allocate(
+    #     page_count=config.hp.context_length // config.block_seq_stride
+    # )
+
+    for block in model.attn_blocks.modules():
+        block.cache_state = torch.concat()
 
     logits = model.prefill(
         tokens=ids,
